refactoring - making Optional in case of not existing the data

+++ src/main/java/org/acme/telemetryservice/domain/dto/query/CoffeeMachineTotalCoffeesMade.java
@@ -8,5 +8,5 @@ public interface CoffeeMachineTotalCoffeesMade {
     @NotNull
     UUID getDeviceId();
 
-    long getAllTimeTotalCoffeesMade();
+    int getAllTimeTotalCoffeesMade();
 }
